convert group-reducer to ts
